+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # v1.1.2
 - Added helper for checksums, good for:
 - Added cache busting support for uris. e.g. "/theme/screen.css?v=HASHHERE445435435ETC"
+- Added support for `Gumdrop#no_render(*paths)`. Treats all matched paths as binary.
 
 # v1.1.1
 - Cleaned up dependencies.